+++ src/data/src/html/templates_dev.rs
@@ -143,7 +143,7 @@ impl HtmlTemplate {
                 };
                 let mut ml = MagicLink::create(
                     user.id.clone(),
-                    RauthyConfig::get().vars.lifetimes.magic_link_pwd_reset as i64,
+                    RauthyConfig::get().vars.lifetimes.magic_link_pwd_reset,
                     usage,
                 )
                 .await?;
